requirements.txt
pandas
matplotlib
seaborn
scikit-learn
sqlalchemy
psycopg2-binary
notebook
python-dotenv
